samplecode/backend-apis/requirements.txt
fastapi
uvicorn
langchain
langchain-community
faiss-cpu
pypdf
openai
python-dotenv
google-generativeai
langchain-google-genai